chore(#694): one Censuses target replaces one-target-per-cluster ClusterACensus
Cluster A's ClusterACensus executable target becomes the cluster-a subcommand of
a shared Censuses target: swift run Censuses cluster-a, swift run Censuses cluster-b
once Cluster B lands, swift run Censuses with no argument lists what is available
instead of failing.

Chosen per #694's option 3, for the reasons the maintainer gave: a target whose
path points directly at a Scripts/repro/<cluster> directory makes renaming that
directory fail the whole manifest with "invalid custom path", one exclude: list
per cluster drifts as new files are dropped in, and Cluster B wants the same
split-box compound fixtures Cluster A already built.

Source moves to Scripts/repro/censuses/ (ClusterA.swift, SharedFixtures.swift,
CensusRunner.swift, main.swift). Cluster A's own directory keeps only its README
and static classifier, neither of which SwiftPM needs to see, so renaming it no
longer touches the manifest at all: confirmed by reproducing the pre-#694 failure
in a disposable worktree at the base commit, then confirming the same rename is
inert on this branch. Cluster A's 45 rows are unchanged, verified byte-identical
against the prior ClusterACensus output.
