+++ Scripts/repro/censuses/ClusterB.swift
@@ -1,0 +1,9 @@
+// Cluster B census (#665): placeholder, filled in next.
+import Foundation
+import OCCTSwift
+
+enum ClusterB {
+    static func run() {
+        print("Cluster B census: not yet implemented.")
+    }
+}
